+++ src/ionic-serve.ts
@@ -10,26 +10,26 @@ import { Project } from './project';
  * Create the ionic serve command
  * @returns string
  */
-export function ionicServe(project: Project, browser: string): string {
+export function ionicServe(project: Project, dontOpenBrowser: boolean): string {
   switch (project.repoType) {
     case MonoRepoType.none:
-      return ionicCLIServe(project);
+      return ionicCLIServe(project, dontOpenBrowser);
     case MonoRepoType.nx:
       return nxServe(project);
     case MonoRepoType.npm:
     case MonoRepoType.folder:
-      return InternalCommand.cwd + ionicCLIServe(project);
+      return InternalCommand.cwd + ionicCLIServe(project, dontOpenBrowser);
     default:
       throw new Error('Unsupported Monorepo type');
   }
 }
 
-function ionicCLIServe(project: Project): string {
+function ionicCLIServe(project: Project, dontOpenBrowser: boolean): string {
   const preop = preflightNPMCheck(project);
   const httpsForWeb = vscode.workspace.getConfiguration('ionic').get('httpsForWeb');
   const previewInEditor = vscode.workspace.getConfiguration('ionic').get('previewInEditor');
   let serveFlags = '';
-  if (previewInEditor || ionicState.webDebugMode) {
+  if (previewInEditor || dontOpenBrowser) {
     serveFlags += ' --no-open';
   }
   if (httpsForWeb) {